temporarily move M4 NVIC to main bus
This is a workaround for the issue of NVIC not being visible
from inside amba_trch memory region. Eventually, we might
need to bring back the amba_trch region and map the NVIC into it.

The difficulty there is that nvic code creates the region
programatically (and associates read/write callbacks to it), and it is
not clear how to map that programmatically-created region into amba_trch
region defined in the device tree.  Creating a new region in the device
tree within amba_trch is easy, but then it doesn't have the callbacks.